Auth Service Moved to api/v1
